Split larch/implement/ship.py (2,816 LOC) into 7 cohesive sibling modules
Resolves #5777. The god-module is split into:
- ship_state.py: state file I/O, validation, atomic patch helpers
- ship_result.py: ShipResult type, error handling, terminal writes, emission
- ship_resume.py: ResumePlan parsing, context hydration
- ship_guidelines.py: architectural-guidelines note handling
- ship_merge.py: merge loop helpers, rebase phases
- ship_pr.py: PR title/body helpers, postmerge phase
- ship_seed.py: initial state seeding CLI

No behavior change. ship.py re-exports the public surface for callers.
Complexity-baseline and ruff.toml updated to reflect new module paths.
